requirements.txt
Flask==3.0.3
google-generativeai==0.7.2
slack_bolt==1.19.1
slack_sdk==3.31.0
google-cloud-aiplatform==1.60.0
google-cloud-documentai==2.20.1
backoff==2.2.1
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.1
scipy==1.14.0
python-dotenv==1.0.1
tenacity==9.0.0
vertexai==1.60.0
